Separate browser/non browser build APIs from Console.cs
